% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment.R
\name{final_payment}
\alias{final_payment}
\title{Total of the principle of final payment}
\usage{
final_payment(a, r, n)
}
\description{
This function is 'Total of the principle of final payment'
a is the principal sum
r is compound interest rate
n is payment period
}
